requirements-py311.txt
customtkinter==5.2.2
python-docx==1.1.2
reportlab==4.2.5

# Force wheel-compatible versions on CPython 3.11 (Windows)
pillow==10.4.0
lxml==5.3.0
